environmentd: detect malformed rows
pgwire had runtime protection against mismatching desc and rows. Extend that to
http and websocket. This would have protected against the panic in the linked
issue. Adding a test for this results in panic without this patch, and an error
output with it. Not adding the test because the underlying issue is getting
fixed which will prevent the test from working.

See #26549