compute: don't assert on current storage collection sinces
Preparatory work for
https://github.com/MaterializeInc/materialize/issues/24845, where we
want to introduce more concurrency to the Coordinator and Controllers.

As the comments in code describe, we cannot assert but there are already
other checks in place that will catch bugs.

Refresher on used terms:

- implied capability: the frontier that is derived from the upper via
  the ReadPolicy. This does not take into account outstanding read
  holds.
- read capabilities: the frontier of outstanding read holds plus the
  implied capability.

The reason we can't assert on the implied capability is that at any
given time it might either be dominated by the absolute ReadPolicy that
the adapter holds or by the "lag writes" policy. We don't know which is
the case and if it's the latter the implied capability will move on when
the upper advances concurrently.

As an alternative, you might think that we can assert on the _read
capabilities_, the frontier of all outstanding read holds (which
includes the read holds that the caller has hopefully installed) and the
implied capability. However, here we also don't know which read holds
dominate the frontier: if it's someone else's read holds the since can
again advance concurrently if/when those are released.

Overall, with the added concurrency that we will introduce over time,
the only thing that we can rely on is read holds that we ourselves put
in place and hold, while we hold them.