tests: always retry in check-shard-tombstone
Preparatory work for
https://github.com/MaterializeInc/materialize/issues/24845, where we
want to introduce more concurrency to the Coordinator and Controllers.

The heuristic of only retrying when there is no more critical handle
registered is not always correct. With the refactored
StorageController/StorageCollections the handle can be expired early but
the append to the empty upper antichain _or_ the downgrade to the empty
since can take a bit longer. This makes the test action more lenient,
giving more time to converge on the expected state.