compute: emit `ReadCapability` responses
This commit makes compute replicas emit `ReadCapability` responses that
track the meet of all input frontiers for each collection. The approach
is similar to how upper frontiers are reported, including making sure
that final responses are emitted for dropped collections and that no
responses are emitted for collections replaced during reconciliation.