catalog: Listen to builtin item updates (#26542)
This commit teaches the in-memory catalog how to listen to builtin
item updates. For now the catalog should only ever see positive changes
to builtin items, and it should only see those changes during startup.
The set and schema of builtin items cannot change in-between version
upgrades. So this change isn't strictly necessary, however it allows us
to initialize the catalog state for builtin items using the same
framework that we use to initialize the catalog state for user items.

Works towards resolving #24844