persist: require a TotalOrder Timestamp for rewrite_ts
This is because we need to be able to verify that the contained data,
after the rewrite forward operation, still respects the new upper. It
turns out that, given the metadata persist currently collects during
batch collection, this is possible for totally ordered times, but it's
known to be _not possible_ for partially ordered times. It is believed
that we could fix this by collecting different metadata in batch
creation (e.g. the join of or an antichain of the original contained
timestamps), but the experience of this bug has shaken our confidence in
our own abilities to reason about partially ordered times and anyway all
the initial uses have totally ordered times.

Closes #26384