RBAC: Fix concurrent role drop check
Previously, RBAC would check that a users role was still valid (i.e.
not concurrently dropped) before doing anything else.
4db9a3cf24b2dc7454a9eeafa9644ab2ca305441 moved this check so that it
happened after collecting the role membership. As a result, it was
possible to try and look up a role in the catalog before checking that
the role exists. This would cause a panic instead of a user error.

This commit fixes the described issue by moving the valid role check
back to being the first thing that RBAC does.

Fixes #26622