doc/user: make fnlist tables readable
Our functions tables have gotten increasingly unreadable over the years
due to support for functions with long signatures and description.
Restore the tables' readability using the same trick as the PostgreSQL
docs to display the function names and descriptions on separate lines of
one column, rather than in two columns.

Co-authored-by: Marta Paes <marta.paes.moreira@gmail.com>