adapter,controller: restrict access to innards of storage controller
Preparatory work for
https://github.com/MaterializeInc/materialize/issues/24845, where we
want to introduce more concurrency to the Coordinator and Controllers.

We restrict access to `CollectionState`, because we plan to move that
out of `StorageController`. But it's good in general to hide this
internal detail of the controller.

Access to `CollectionState` is replaced by more targeted methods, this
means there's now more methods but they return "smaller" things.
Additionally, the new methods return owned data rather than references
into controller state. This would not be necessary for how the
controller and calling code work now, but future refactoring will make
it so that state inside the controller is behind a shareable mutex, so
we start handing out owned data now in preparation for that.