compute: emit `input_frontier`s in `Frontiers` responses
This commit makes compute replicas emit `input_frontier` updates in
`Frontiers` responses that track the meet of all input frontiers for
each collection. The approach is similar to how upper frontiers are
reported, including making sure that final responses are emitted for
dropped collections and that no responses are emitted for collections
replaced during reconciliation.