storage: factor StorageCollections out of StorageController
Preparatory work for #24845, where we want to introduce more concurrency
to the Coordinator and Controllers.

The considerations/design are described in
[doc/developer/design/20240117_decoupled_storage_controller.md](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/20240117_decoupled_storage_controller.md)

This is an intermediate step where we factor a `StorageCollections` out
of `StorageController`, and let the `StorageController` use it's
interface instead of holding collections state/sinces itself. Note that
there are now a lot of methods in `StorageController` that pass through
to `StorageCollections`. A large part of this can be removed once we
change usage sites to use a `StorageCollections` directly.

One of the next steps is to change usage sites of `StorageController` to
use their own handle to a `StorageCollections`, bypassing the
`StorageController` for query-processing (PEEKS, SUBSCRIBE, etc.) code
paths. This will let us introduce more concurrency in the Coordinator
and do less work in the main Coordinator loop.

The important parts in this change are:

- `StorageCollections::new` and `StorageController::new`: these closely
  mirror each other.
- `StorageCollections::create_collections` and
  `StorageController::create_collections`, ditto!
- A lot of the rest is "boilerplate", passing through calls to the
  internal `StorageCollections`, and using `StorageCollections` from the
  controller instead of using owned state.
- The last interesting thing to look at is the new `BackgroundTask`: it
  takes over the work that `persist_handles::PersistReadWorker` was
  doing before plus it continually listens for upper changes and
  forwards the since frontier of collections/their since handles.