columnar: Refactor `PartDecorder` (#26605)
This PR refactored `PartDecoder` to have owned values, instead of
references to the columns in a `Part`. This is required for efficiently
working with structured columnar data.

Note: this PR originally included a change for `PartEncoder` as well,
but it was split out into a separate PR.

### Motivation

A `Part` is our structured columnar representation of data. Currently to
decode a `Part` we take references to the contained columns and downcast
them from a `dyn Array` trait object, to a concrete column type. This
downcasting is relatively expensive so we want to do it as infrequently
as possible, ideally once per-`Part`.

The problem is when iterating over a `Part` (i.e. with `FetchedPart`) we
own the underlying data, and the lexical scope for which we would be
able to borrow is a single iteration. This means we would need to create
a `PartDecoder` for every `Row` we decode, which means repeatedly
downcasting our trait objects, something we need to avoid for
performance reasons. The problem we're running into is a classical Rust
issue, to prevent downcasting on every iteration what we'd need is
something like:

```
struct FetchedPart {
  part: Part,
  decoder: PartDecoder<'self>,
  // ...
}
```

where the lifetime of the `PartDecoder` is tied to `FetchedPart`, but
this requires self references and isn't easily done in Rust. (Note: the
`Pin` type exists to achieve self-referential structs, but I wanted to
avoid using that here).

#### Performance

This is a performance sensitive code path, but there shouldn't be any
noticeable change with this refactor. Instead of passing `&T` we now
pass `Arc<T>`. All of these types are the same size (one word) so they
should all get passed via registers in function calls. There is a bit of
overhead with de-referencing an `Arc`, you need to do some pointer
arithmetic to get the address of the inner value, but that is simply
something we can't avoid.

### Tips for reviewer

The PR is split into three separate commits which can be reviewed
independently:

1. Refactor of `PartDecoder`
2. Refactor of benchmarks

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a