storage: add explicit ReadHolds API
Preparatory work for #24845, where we want to introduce more concurrency
to the Coordinator and Controllers.

This introduces an API with an explicit `ReadHolds` token that holds the
frontier of a collection while it is "live". In the long run, this will
replace low-level handling of read holds via `update_read_capabilities`.

Before, acquiring a read hold worked like this:

1. look at collection since, using `collection_frontiers`
2. assume and pray that frontiers don't move on concurrently
3. put in place a read hold using `update_read_capabilities`

This works, because there is almost no concurrency between the
adapter/coordinator/controller, except for select moments. It won't work
in the future, though, when we introduce more concurrency. Especially
because frontiers will move on concurrently.

With the new API, acquiring a read hold is one atomic request on the
controller.