storage/{pg,mysql}: serialize snapshot and replication phases
This PR makes it so pg and mysql sources don't start reading their
replication streams before finishing with the snapshot. We have observed
that doing so leads to a lot of updates being emitted into the dataflow
that can't be fully forgotten until the frontier advances. If the
snapshot is large that can be a long time and a lot of data.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>