sql-parser: correct parsing of `interval` keyword
To match PostgreSQL, treat `interval` as a non-reserved keyword. This
allows it to be used as a column name without quoting.

This actually was very nearly already the case. When we implemented
general typed string literals (77811560), we got this behavior "for
free" for intervals, but failed to remove the *old* implementation for
parsing interval values. So all we need to do in this commit is remove
that old implementation, to allow the parsing of intervals to be
entirely handled by the new general typed string literal parser, which
correctly allows type names to *not* be treated as reserved keywords.