[persist] Split up runs that might not truly be consolidated (#26409)
In the distant past, we wrote some >1-length runs that aren't actually
consolidated: due to bugs or to ordering changes. This is annoying
because the Consolidator requires its inputs to be consolidated to
produce consolidated output. So: we have code that will re-sort parts
that it discovers were written by old versions of Materialize.

However, even if all parts are consolidated, runs might not be! (Oops!)
So it's safest to treat runs written using these old versions as though
they were a bunch of single-part runs instead.

### Motivation

Known bug: #26170 

This assert has only fired once ever in production, and we didn't have
enough debugging info enabled at the time to know for sure that this is
the exact issue we hit.

### Tips for reviewer

I started work on a unit test for this, but it is tedious to write
because it requires generating batches that aren't correct, which
involves a fair bit of low-level code. Let me know if that seems
worthwhile!

I've also added some additional context to the log, with hopefully
enough information to start an investigation if this assert fires again.
Happy to take suggestions on format / content / etc.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- <!-- Add release notes here or explicitly state that there are no
user-facing behavior changes. -->