compute: suspend index imports
This commit makes compute replicas suspend index imports until the
dataflow start signal is fired (i.e. the dataflow is scheduled).
Suspension works by adding an operator after the `import_frontier`
operator that holds back any emitted data until it observes the start
signal. The data needs to be buffered in the time between, but it
consists only of references to batches, so this should be efficient.