catalog: Add sub-operations to transactions (#26781)
This commit adds sub-operations to catalog transactions. This allows
callers to logically group operations within a catalog transaction. This
feature will be useful when the catalog is structured to listen to
durable changes, because it will allow the catalog to extract durable
changes for a single logical operation and apply just those changes to
it's in-memory state.

Works towards resolving #24844