compute: `Correction` since support
This commit adds a `since` frontier to the `Correction` data structure,
which specifies times that clients still reserve the ability to read.
The `since` is used to aggressively advance both contained and inserted
times, so contained updates can be consolidated as much as possible.

The new `since` is not used by the existing MV sink implementation but
will be used by the new one, to avoid inserting updates at times that we
don't care about anymore.

This commit also adds an `insert_negated` method which encapsulates
adding updates read back from persist that should have their diffs
negated.