compute: core persist_sink_v2 implementation
This commit adds the core logic of the new compute persist_sink. Details
like logging and metrics are added in later commits.

persist_sink_v2 aims to be a more readable implementation of the
self-correcting compute persist_sink that avoids unnecessary work as
much as possible. See the module docs of
`mz_compute::sink::persist_sink_v2` for details.