compute: core MV sink v2 implementation
This commit adds the core logic of the new MV sink. Details like logging
and read-only mode are added in later commits.

The new MV sink aims to be a more readable implementation of the
self-correcting persist sink that avoids unnecessary work as much as
possible. See the module docs of
`mz_compute::sink::materialized_view_v2` for details.