adapter: error for slow coord messages
We decided to move to caring about individual slow messages because we largely have control over how long they take to execute. We want to have an SLO/SLI around these. Decide on 30s as the initial starting goal and send to entry when they occur.

Demote the busy log to a warning. We care less about this because many messages is workload dependent and thus can vary a lot by user. We initially will have a lower importance to this metric until the slow individual messages metric is improved.

Fixes #26861