adapter: instrument cluster readiness
One of the cloud team's SLIs this quarter is the amount of time between
the commit of a `CREATE CLUSTER` or `ALTER CLUSTER` command and when the
cluster becomes ready. So, add a Segment event every time a cluster
becomes ready or becomes not ready.

It will then be straightforward to write an analytics query that joins
these new events against the existing "Cluster Replica Created" events
to determine the cluster startup delay. (Under the hood, an `ALTER
CLUSTER` is converted into a sequence of replica drop/creation events.)

Touches MaterializeInc/cloud#9311.