segment: permit setting event timestamp
Allow calls to `track` to set the event timestamp. This is useful in
situations where there is a large delay between when the event occurs
and when the event is reported to Segment.

To avoid adding yet another parameter to the `track` method, this commit
introduces an options struct, so that optional parameters that are not
relevant at a particular callsite can be neatly elided via a defaults
splat (`..Default::default`).