sql: improve coercion of record literals
Previously, the following query would fail to plan in Materialize:

    SELECT LIST[('a', 1), ('b', 2)]

The issue was in our selection of a best common type for the list
elements. `guess_best_common_type` would assess that there
were no list elements with a known (i.e., coerced) type and fall back to
using `text`. The query would then fail to plan, since the record
literals are not coercible to `text`.

If all list elements are uncoerced record literals, the best common type
is some sort of record type. This commit adjusts
`guess_best_common_type` accordingly. The new implementation detects
this situation and constructs a record type as the best common type,
where the type of each field is the result of guessing the best common
type for that field.

This commit also hoists planning of `IN` lists (e.g., `x IN (1, 2)`)
involving record types into an AST transformation. This is necessary to
preserve the existing support for applying the `IN` list operator to
expressions of type record. The old `plan_in_lists` function, when
combined with the new coercible record literal implementation, no longer
properly handles `IN` lists with expressions of type record. The good
news is that the new approach exactly matches the PostgreSQL
implementation.

The specific motivation for this change is to make it possible to call
`map_build` with anonymous record types. The following query now plans
correctly:

    SELECT map_build(LIST[('a', 1), ('b', 2)])

Previously, calling `map_build` required constructing a custom record
type of the appropriate shape and explicitly casting the list elements
to that type:

    CREATE TYPE header AS (k text, v int);
    SELECT map_build(LIST[('a', 1), ('b', 2)]::header list)