sql: replcae AstInfo::ColumnName with AstInfo::ColumnReference
When buidling Visitors, if a visitor only defined a visitation to
AstInfo::ItemName, it would not be visited if it was inside an
AstInfo::ColumnName. The code that generates our visitors assumes
that AstInfo's associated types are "terminal," and recursing
into them requires an explicit implementation.

This seems like a footgun, so instead move the footgun to a place
that's more obvious and treat `ColumName` as a regular AST element,
and then make both of its fields into AstInfo associated types,
T::Itemname and T::ColumnReference, respectively.