persist-txn: Fix since_ts invariant (#26969)
The `TxnsCacheState` has an invariant, that is written in the
`compact_to()` method but not the struct definition, that states
`self.since_ts <= self.progress_exclusive`. Previously, when
initializing the cache, `progress_exclusive` was always initialized to
`T::minimum()` while `since_ts` was always initialized to the since of
the txns shard. As a result, almost all caches were initialized in a
broken state. This led to the confusing scenario where the cache may be
learning about new entries that should have already been compacted.

This commit fixes the issue by always initializing the cache s.t. both
the initial `since_ts` and the initial progress_exclusive` are advanced
to the since of the txn shard.

Touches #26893