storage/pg: adaptively pick the feedback interval value
This PR reverts #26818 and makes the feedback interval value used by our
replication client depend on the timeout value on the server instead of
hardcoding it to 30 seconds.

This allows db administrators to set the `wal_sender_timeout` to any
value they think is right without risk of their source timing out.

Closes #26813

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>