storage/upsert: thin out data before exchange
This PR renders an operator that discards updates that are known to not
affect the outcome of upsert in a streaming fashion. For each distinct
`(key, time)` in the input it emits the `value` with the highest
`from_time`. Its purpose is to thin out data as much as possible before
exchanging them across workers which is benefitial for flow control in
the presence of skew.