[persist] Replace the lease returner with an Arc-based count (#27021)
This swaps out Persist's leasing mechanism: instead of explicitly
returning leases to the reader that issued them, we generate a lease
token that keeps the lease alive as long as it's held.

This works transparently for most `LeasedBatchPart` usage... the only
difference is that there's no need to return the part explicitly.
However, we need some other way to do lease management in the shard
source, since we still can't send leases across process boundaries. This
code makes serializing a part also spit out the lease, and makes the
source hold on to the lease until a frontier advances.

### Motivation

This is useful for #26939 - eg. it would be convenient if I could turn a
single `LeasedBatchPart` (which actually represents a list of hollow
parts stored elsewhere) into a bunch of `LeasedBatchPart`s representing
the individual sub-parts, potentially in code like the `Consolidator`
which doesn't have direct access to the handle, and being able to
explicitly manipulate and clone leases helps with that.

(OTOH, I think this may be worthwhile even without #26939 on deck --
especially in the common case of leased parts, it seems to make things
easier to manage. Happy to break that down if you don't immediately
agree however!)

### Tips for reviewer

There are a bunch of possible followups one might imagine, but I've
tried to keep this to just the core change and leave everything else
alone.