Efficient copy of DatumSeq
Previously, copying the contents of a DatumSeq would require iterating the
datums in the DatumSeq and copying each item individually. With this
change, we exploit the fact that the row encoding in DatumSeq is a correct
row encoding, which allows us to copy the contents directly, without
going through their datum representation.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>