Support container builders for flat_map variants
This change adds support for specifying the container builder to the
various variants of `flat_map` and adjusts the call sites. Most use the
default capacity-based builder, but some switch to the consolidating
variant. I selected places that apply scalar expressions to rows for the
consolidating variant, which have the chance to change rows sufficiently
that it enables further consolidation.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>