Create historical versions of source statistics built in views
Due to an issue with latency and high CPU usage, we create a non-historical index + view on top of mz_source_statistics such that Console's source lists efficiently query on it. The latency and high CPU usage is related to us retaining history, causing the number of updates in the index to be large.