Consolidate on references instead of owned data
The current implementation converts a `TupleRef` iterator into a vector and
consolidates that. This might cause extra work if the data consolidates
well because it needs to convert the whole input into owned data.

With this change, we first collect the references into a vector,
consolidate this, and then convert into an owned allocation.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>