persist-txn: Fix registered_at idempotency (#26930)
Previously, the `registered_at` and `all_registered_at` functions in
`TxnCacheState` allowed callers to ask about a time in the future. If a
timestamp in the future was passed as an argument, then the functions
would return an answer as if nothing changed from now to that timestamp.
If something did in fact change and the same timestamp was passed as an
argument, then the answer returned would change.

As it turns out, almost all callers of `registered_at` and
`all_registered_at` actually wanted to know information as of the
current progress of the cache, not as of some future timestamp. This
commit removes the timestamp argument from these functions and always
returns a result as of the current progress. There is prior art here in
the `unapplied` function, which does the same thing.

The one caller of `registered_at` that was interested in some time in
the past, was updated to use a different function.

We did keep `all_registered_at` for the one caller that was interested
in all the data shards registered at some point in the past, but we
added an assert to reject times in the future.

Fixes #26903