sql-parser: correct parsing of `SELECT AS OF ...`
`SELECT AS OF ...` is a valid SQL query that produces a zero-column,
zero-row relation at the specified as-of time. Not very useful, but a
semantically valid thing to do nonetheless.

@aalexandrov noticed in #27167 that this query was being misparsed in a
way which did not roundtrip, which could result in a broken catalog if
such a query were used as the definition of a materialized view.

Fix #27167.