storage: store the cluster running ingestions and their exports
We want to notify clusters if ingestion exports get dropped. The natural place
to find that information is by looking up the export's ingestion in the
controller's set of running collections.

However, PostgreSQL and MySQL sources have no real need to install a dependency
on the primary ingestion's collection, which means that ingestion can be dropped
from the in-memory state before its exports. This occurs when e.g. the ingestion
is dropped w/ `CASCADE` but the exports have active read holds placed against
them.

To remedy this tension, store the cluster running the ingestion and the export
in the collection metadata. This is a temporary workaround until #20208 lands,
in which case we can always just look up the ingestion's cluster ID (ingestion
exports will rely on the "ingestion" directly, because what is now the remap
collection will become the ingestion's collection).