compute: avoid sending unexpected `Frontiers` responses
This commit fixes a bug in `report_dropped_collections`, which tried to
ensure that no final `Frontiers` responses were emitted for subscribes,
but failed because the `is_subscribe` method checked for the existence
of a sink token, which was already dropped by `drop_collection`.

The fix is to replace the `is_subscribe` method with a flag in
`CollectionState` that explicitly remembers whether a collection is a
subscribe or not.

Apart from subscribes, copy-tos have the same behavior of not producing
`Frontiers` responses. This is also made explicit now in the code.