adapter: disallow mz_now in all places of RTW statements
Previously we (incorrectly) allowed mz_now in UPDATE, DELETE, RETURNING, and
INSERT if in a dependent view. Fix all of these. The rationale for disallowing
it is that mz_now will produce a time during the Read part of RTW, which will
differ from the timestamp of the committed data during the Write part.

Fixes #26297