storage: allow configuring load generator as of and up to
Add new `AS OF` and `UP TO` options to all simple (non-KV) load
generators. The former option controls the size of the initial snapshot,
while the latter option controls the maximum amount of data generated by
the load generator. Both are measured in number of ticks.

For example, with the default tick interval of 1s, setting `AS OF` to
300 and `UP TO` to 420 will create a load generator that starts with 5m
worth of data and runs for 2m before ceasing.

This will be useful in the console's quickstart, which falls a little
flat at the moment because building an index on a half dozen rows
doesn't demonstrate how much faster the index makes queries. With these
new load generator options, the console quickstart will be able to use a
load generator that starts off with much more data and makes for a much
better demo.

These options also turn out to be useful in tests, as they allow
creating load generators that produce a fixed amount of output against
which assertions can be easily written.