storage: rm misguided soft assert when altering SourceDesc
I added an assert that all SourceDesc changes should actually
change the SourceDesc. However, I had overlooked the impossible
to detect case where:
1. User creates PG source foo with tables a, b.
2. User drops subsource b. This does not update the SourceDesc.
3. User re-adds subsource b without having changed its schema.

In this case, the actual source desc will not have changed whatsoever
and spuriously trip the assert. Unclear how to really detect this case,
and in light of this use, the assert is of questionable value.