catalog: Move state update sorting
This commit is mostly code movement and code organization. Previously,
`apply_updates_for_bootstrap` relied on callers to sort the input in
the correct order. That made it easy to misuse
`apply_updates_for_bootstrap`, which was fine because there was only a
single caller. Soon, we will be adding more callers, which would
require us to duplicate the sorting code for all callers. Instead, this
commit moves the sorting code to inside `apply_updates_for_bootstrap`.
Callers can pass in input in any order and it will be sorted for them.

Works towards resolving #24844