persist-client: fix race in sink peak metrics update
The `SinkMetrics::update_sink_correction_peak_metrics` method assumed
"quiescence", i.e. the temporary ceasing of updates to sink metrics
while it was executing. While the method was correct under that
assumption, the calling code unfortunately was not able to uphold it.
Specifically, `step_or_park` is invoked per worker, so when
`update_sink_correction_peak_metrics` is invoked it can only assume that
the current worker won't update the sink metrics, but it cannot assume
anything about other workers.

As a result of the violated assumption, "Negative aggregate length for
persist sink correction" could be logged even when nothing was actually
wrong with the persist sink implementation.

Stopping all workers at the same time regularly to update the sink
metrics is not an option, so instead of trying to make the calling code
uphold the quiescence assumption, this commit instead modifies
`update_sink_correction_peak_metrics` to not require quiescence anymore.
In essence this just removes the error logging, but the code is also
restructured to improve readability.