storage: eagerly clear read holds when dropping collections
Fixes #27304

My recent change that factors a `StorageCollections` out of the
`StorageController`
(https://github.com/MaterializeInc/materialize/pull/27057) made an
existing bug more problematic. See below!

Before the mentioned change, this could happen:

1. create cluster c
2. create source/sink on cluster c
3. do a DROP CLUSTER c cascade
4. cluster processes are killed before they get a chance to send back
   DroppedIds messages
5. controller does not clean out state about that collection (until the
   next restart)

With my refactor, this would happen, which manifests in the observed bug:

1. create cluster c
2. create source/sink on cluster c, **this acquires a read hold at
   StorageCollections and stores it in that collection's state**
3. do a DROP CLUSTER c cascade
4. cluster processes are killed before they get a chance to send back
   DroppedIds messages
5. controller does not clean out state about that collection (until the
   next restart)
6. **the read hold is never released**, meaning StorageCollections does
   not clean out some of its state and we still report frontiers for
   these "active collections"

Due to reasons (tm) my previous changes did make it so that we only drop
read holds when getting a DroppedIds message: the previous code would
only attempt shard finalization after getting a DroppedIds, and
StorageCollections starts to attempt shard finalization when all read
holds have been dropped. This preserved the previous behavior of
StorageController.

With this here change, we allow eagerly dropping the read holds
(advancing their since to the empty frontier), which we previously did
not allow, on purpose. This makes is so that we correctly drop their
state, and no longer report their frontiers. But it also makes it so
that we attempt shard finalization slightly earlier. I think that is
okay, though.